Update BluetoothSingleton Mehtod. Add Mehtod connectGatt to all ESP32 Commands

+++ app/src/main/java/com/example/cocktailmachine/ui/fillAnimation/FillAnimation.java
@@ -85,7 +85,7 @@ public class FillAnimation extends AppCompatActivity {
                 Bitmap image = null;
                 if(true){
                     try {
-                        image = BildgeneratorGlas.bildgenerationGlas(context,recipe,(float)animatedValue);
+                        image = BildgeneratorGlas.bildgenerationGlas(context,recipe, animatedValue);
                     } catch (TooManyTimesSettedIngredientEcxception | NoSuchIngredientSettedException e) {
                         e.printStackTrace();
                     }
